+++ src/Domain/Models/Entities/QueryBuilder.php
@@ -3,6 +3,7 @@
 namespace Soyhuce\NextIdeHelper\Domain\Models\Entities;
 
 use Illuminate\Database\Eloquent\Builder as EloquentBuilder;
+use Illuminate\Support\Collection;
 use Soyhuce\NextIdeHelper\Support\Type;
 
 class QueryBuilder
@@ -11,14 +12,24 @@ class QueryBuilder
 
     public string $filePath;
 
+    public Collection $extras;
+
     public function __construct(string $fqcn, string $filePath)
     {
         $this->fqcn = Type::qualify($fqcn);
         $this->filePath = $filePath;
+        $this->extras = collect();
     }
 
     public function isBuiltIn()
     {
         return $this->fqcn === '\\' . EloquentBuilder::class;
+    }
+
+    public function addExtras(string $extra): self
+    {
+        $this->extras->add($extra);
+
+        return $this;
     }
 }